# std.Io.Reader: encourage inlining hot buffer check