Resolves: #24424

The benchmark in that PR is now *significantly* faster, even for the `take(10)` case.

Before:
```sh-session
$ ./repro fast
Time: 860ms
$ ./repro slow
Time: 1479ms
```

After:
```sh-session
$ ./repro fast
Time: 349ms
$ ./repro slow
Time: 358ms
```

The single `take` does still consistently *slightly* beat out the double `takeInt`. That makes sense; the second approach has to do a tiny bit more work (an extra check on the buffer length). However, the difference is ~2-5% instead of ~70%, and even the "slow" path is now 60% faster than the "fast" path *previously* was.